requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
datasets>=2.14.0
evaluate>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
biopython>=1.81

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.10.0

# Experiment tracking
wandb>=0.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Utilities
python-dotenv>=1.0.0